hardening

+++ genomekey_deploy/fab/aws.py
@@ -10,11 +10,12 @@ from genomekey_deploy.util import apt_update
 #
 
 __author__ = 'erik'
+GENOME_KEY_USER = 'genomekey'
 
 
 @task
 def init_node():
-    with hide('output'):
+    with settings(user='root'), hide('output'):
         # note can make an AMI to avoid doing this
         # TODO get rid of this pastebin
         run('wget "http://pastebin.com/raw.php?i=uzhrtg5M" -O /etc/apt/sources.list')
@@ -28,41 +29,34 @@ def init_node():
                 'echo oracle-java7-installer shared/accepted-oracle-license-v1-1 select true | /usr/bin/debconf-set-selections')
             run('apt-get install oracle-java7-installer oracle-java7-set-default -y')
 
-        run('mkdir -p /mnt/genomekey/scratch')
-        run('chown genomekey:genomekey /mnt/genomekey/scratch')
-
-        # if not files.exists('/scratch'):
-        # run('mkdir -p /mnt/genomekey/scratch')
-        #     run('ln -s /mnt/genomekey/scratch /scratch')
-        #     run('chown genomekey:genomekey /mnt/genomekey/scratch')
+        run('chown -R genomekey:genomekey /genomekey') # TODO fix the perms in the ami
 
         with settings(user='genomekey'):
-            setup_aws_cli()
-            # mount_genomekey_share()
-            # if sync_genomekey_share:
+            run('mkdir -p /mnt/genomekey/scratch')
 
-        sync_genomekey_share()
+            sync_genomekey_share()
 
 
 @task
 def init_master():
-    with hide('output'):
+    with settings(user='root'), hide('output'):
         # update apt-list, starcluster AMI is out_dir-of-date
         run('apt-get install graphviz graphviz-dev mbuffer -y')
         run('pip install awscli')
 
-        with settings(user='genomekey'):
+        with settings(user=GENOME_KEY_USER):
             run('mkdir -p /home/genomekey/analysis')
 
         # For ipython notebook.  Do this last user can get started.  Installing pandas is slow.
         run('pip install "ipython[notebook]" -I')
 
-        with settings(user='genomekey'):
+        with settings(user=GENOME_KEY_USER):
             files.append('~/.bashrc', ['export SGE_ROOT=/opt/sge6',
                                        'export PATH=$PATH:/opt/sge6/bin/linux-x64:$HOME/bin', ])
+            setup_aws_cli()
 
 
-# @task
+# @taska
 # def mount_genomekey_share(mount_path='/mnt/genomekey/share_yas3fs'):
 # with aws_credentials(), settings(user='root'):
 #         run('pip install yas3fs')
@@ -74,21 +68,21 @@ def init_master():
 #             chmod_opt(os.path.join(mount_path, 'opt'))
 #             # '--topic arn:aws:sns:us-west-2:502193849168:genomekey-data --new-queue --mkdir')
 
-@task
-def sync_genomekey_share(user='genomekey'):
-    with settings(user='root'):
+def sync_genomekey_share(user=GENOME_KEY_USER):
+    """
+    Requires that setup_aws_cli() has already been called
+    """
+    # with settings(user='root'):
         # TODO change the AMI and delete this?  This runs instantly so not a big deal
         # run('chown -R genomekey:genomekey /genomekey')
-        pass
-    with settings(user=user):
-        with hide('output'):
-            setup_aws_cli()
-            if files.exists('/genomekey/share'):
-                run('ln -s /genomekey/share /mnt/genomekey/share')
-            else:
-                run('mkdir -p /mnt/genomekey/share')
-            run('aws s3 sync s3://genomekey-data /mnt/genomekey/share')
-            chmod_opt('/mnt/genomekey/share/opt')
+
+    with hide('output'):
+        if files.exists('/genomekey/share') and not files.exists('/mnt/genomekey/share'):
+            run('ln -s /genomekey/share /mnt/genomekey/share')
+        else:
+            run('mkdir -p /mnt/genomekey/share')
+        run('aws s3 sync s3://genomekey-data /mnt/genomekey/share')
+        chmod_opt('/mnt/genomekey/share/opt')
 
 
 def chmod_opt(opt_path):
@@ -106,7 +100,7 @@ def chmod_opt(opt_path):
 
 
 @task
-def setup_aws_cli(user='genomekey', overwrite=False):
+def setup_aws_cli(user=GENOME_KEY_USER, overwrite=False):
     with settings(user=user):
         if overwrite or not files.exists('~/.aws/config'):
             run('mkdir -p ~/.aws')
